Inactive settings no longer affect settings string
`settings.sav` still holds all setting values, including preferences
that were unnecessary in the last generation. This prevents the
preferences from being lost if the setting is reactivated.

* Add hidden setting `inactive_settings` to track irrelevant settings.
* Add `get_setting_info` to retrieve an individual setting info.
* Change `get_settings_string` to create a string deterministically in
the case of extra, irrelevant setting values such as `trials`. This is
done by substituting the inactive value with the setting info's default
attribute.
* `check_dependency` now updates `inactive_settings` each time it is
called for a setting.