Ignore disabled options in setting string (Issue #206)
The setting strings are different, but they will still produce the same seed. This allows options to keep they state when disabling. that being said, people will almost always just copy setting strings in so this is relatively moot. I'm unsure if this behavior is honestly more expected, but I'm adding this is as mainly a UX reaction since people seem to think it should be this way.