Fix Ganon towerless death warp and health fill
Fix the no-tower-collapse Ganon death warp. Change the immediate
hardcoded halfword in code that sets the entrance in death handling
code.

Add a post scene check that detect when entering the Ganon fight
with towerless and give a full health and magic refill to match the
one Ganon gives you in vanilla after the kill. Also matches vanilla
with the save warp which is supposed to give full health and magic
from top of tower collapse re-attempt.