Zelda's Letter Opens Mask Shop / Complete Mask Quest Option
If Open Kakariko Gate is enabled, obtaining Zelda's Letter will open the mask shop without showing the letter to the Kakariko guard.

A new option has been added to complete the mask quest from the start. If enabled, once the mask shop is opened, all masks are available to borrow.