[GUI] Restore Readme button
The old GUI had a link to launch the readme wiki page. This was lost during the transition to the electron angular gui. This restores that button.

The button should also appear on the website, however I'm not entirely sure how it will look in that case as I don't have the means to test it.

I also didn't try to do any sort of error handling. I believe the browser that is launched (or the new tab in the case of the website (hopefully)) handles that in a user-friendly enough manner already.