Fix start with bottles in plentiful item pool
Starting with 3 bottles in plentiful would cause an exception due to
not seeing a third bottle to remove as Bottle with Letter does not count
as a normal bottle. If not starting with the Bottle with Letter the 
count of bottles started with is reduced by 1 to compensate. This is
done so that there are still 2 Bottle with Letter placed in the world as
is expected of the plentiful item pool. This may confuse players who
wish to start with 3 bottles, however I believe this behavior is more
desirable than an exception, or obtaining a 5th bottle.

Starting with Ruto's Letter in plentiful would start with 1 Bottle with 
Letter and place one in the world. Starting with Ruto's Letter will now 
start with letter and place a random bottle in place of the second one.