Order locations in spoiler log sensibly
Originally the order of locations in the log was based on the order they appeared in the logic files. Since this was not a sensible ordering I decided to reorder them in a manner I deem practical. They are ordered as follows:

1) Dungeon rewards and songs first.
2) All locations are grouped by location (first overworld, then dungeons).
3) The overworld checks follow a route that is close to  the intended vanilla route. Dungeons ditto, but with the mini-dungeons at the end.
4) Within a region the order is gift/freestanding/chest -> scrubs -> cows -> GS -> shops
5) Within dungeons the order follows what I believe is considered the most common/optimal routing in glitchless settings
6) In overworld child checks come before adult checks, and they also follow "optimal" routing (e.g. for night kak skulls you don't go from the watchtower to the skull house to the skull near the DMT entrance)

This sorting is achieved by comparing the list of locations in Spoiler.py with a large list that has been custom sorted by me. There are undoubtedly more efficient and elegant ways of implementing this, which is why I am submitting this as a draft initially. My idea was to add a "sorting index" to the locations in LocationList.py, but this has the disadvantage of making reordering the locations later (e.g. if more locations are added) a drag. I would love to receive suggestions on how to proceed.