Validate that the Light Arrows referenced by Ganondorf is reachable.
Previously, if there were multiple Light Arrows in the item pool and All Locations Reachable was set to off, Ganondorf could have hinted to an unreachable location. Additionally, this shuffles the locations prior to checking them to make sure it is random which reachable location is hinted if there are more than one.