Move logic rules and ALR to the top of the log
These are arguably two of the most important settings the spoiler log can display, so it makes more sense to push them to the top, above almost everything else, rather than being buried in the middle.

This puts these two settings on the fourth and fifth lines of the settings list in the spoiler log, between random settings and closed forest. It *shouldn't* affect anything else.