Spirit Logic Rewrite
I've attempted to rewrite the logic in a simpler way that makes it more clear what each section is doing. And I tried to write a couple comments to explain the reasoning for why the logic in this file is the way that it is. In theory I have not changed the way that the logic actually functions.

Note that it's been some months since I last tested this file. It probably still works, but I suppose it could be possible that something's changed to make it incompatible.