Update weirdshot crash protection to correct value
Recently zfg posted a report indicating that the crash was still happening. After substantial testing, it was found that the crash happened as low as 0x90, even with BOoT.

As I suspect the original writeup just had a typo, I'm changing it to 0x88. Although 0x89 was found to be safe, I'm adding a slight buffer just in case it crashes on higher blue values (which is hypothesized to be possible at high values, such as XX89FF).

I'm also slightly expanding the comment to indicate that this change only takes place without glitchless logic for completion's sake, even though that's *fairly* obvious from the code itself.