Potential route to fix issues revealed by new armips ASM/Makefile: Have the compiler not put zero initialized variables into the .sbss or .bss sections of the elf binary.
ASM/build.py: Error out of the compile if a symbol is
located in either the .bss or .sbss section of the
binary. armips will place these in the binary in
places where they will cause a crash. Both methods
mention (mainly the __attribute__ method) are used in
fixes to the C code in this branch.

C/*: These files have been modified so that nothing
in them gets placed into the problematic sections of
the binary. Both initializing normal variables to 0
and changing the attributes of pointers and structs
are done, so the latter can be referenced as an
example in new code that necessitates it.

I have tested this with the both the old armips and
the latest updated one. Both seem to work. I'm not
currently including the compiled files, but if
someone desires they are added to make testing easier
(which is probably a good idea) I can do so.